requirements_core.txt
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
deepgram-sdk==3.2.7
python-decouple==3.8
ffmpeg-python==0.2.0
requests==2.31.0
werkzeug==3.0.1
spacy>=3.7.0
nltk>=3.8.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
textblob>=0.17.0
python-multipart==0.0.6
verboselogs==1.7